Apply reviewer's comment - implement AbstractAPIController
